Updated the code to fix a segfault issue.
If the user types in "a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a", it used to cause a segfault. Now it doesn't.